+++ scripts/exrtofs_glo_ncoda_hycom_var.sh
@@ -144,21 +144,21 @@ err=$?; export err ; err_chk
 echo " error from rtofs_ncoda_post=",$err
 
 #   rename local files
-mv fort.32 $log_dir/hycom_var.$ddtg.rej
-mv fort.33 $log_dir/hycom_var.$ddtg.prf
-mv fort.34 $log_dir/hycom_var.$ddtg.gpt
-mv fort.36 $log_dir/hycom_var.$ddtg.mvo
-mv fort.37 $log_dir/hycom_var.$ddtg.drc
-mv fort.38 $log_dir/hycom_var.$ddtg.lyp
-mv fort.39 $log_dir/hycom_var.$ddtg.fix
-#mv fort.40 $log_dir/hycom_var.$ddtg.sus
-mv fort.41 $log_dir/hycom_var.$ddtg.dup
-mv fort.42 $log_dir/hycom_var.$ddtg.ssh
-mv fort.52 $log_dir/hycom_var.$ddtg.sal
-mv fort.67 $log_dir/hycom_var.$ddtg.obs
-mv fort.68 $log_dir/hycom_var.$ddtg.grd
-mv fort.69 $log_dir/hycom_var.$ddtg.via
-mv fort.88 $log_dir/hycom_var.$ddtg.dbg
+[[ -f fort.32 ]] && mv fort.32 $log_dir/hycom_var.$ddtg.rej
+[[ -f fort.33 ]] && mv fort.33 $log_dir/hycom_var.$ddtg.prf
+[[ -f fort.34 ]] && mv fort.34 $log_dir/hycom_var.$ddtg.gpt
+[[ -f fort.36 ]] && mv fort.36 $log_dir/hycom_var.$ddtg.mvo
+[[ -f fort.37 ]] && mv fort.37 $log_dir/hycom_var.$ddtg.drc
+[[ -f fort.38 ]] && mv fort.38 $log_dir/hycom_var.$ddtg.lyp
+[[ -f fort.39 ]] && mv fort.39 $log_dir/hycom_var.$ddtg.fix
+[[ -f fort.40 ]] && mv fort.40 $log_dir/hycom_var.$ddtg.sus
+[[ -f fort.41 ]] && mv fort.41 $log_dir/hycom_var.$ddtg.dup
+[[ -f fort.42 ]] && mv fort.42 $log_dir/hycom_var.$ddtg.ssh
+[[ -f fort.52 ]] && mv fort.52 $log_dir/hycom_var.$ddtg.sal
+[[ -f fort.67 ]] && mv fort.67 $log_dir/hycom_var.$ddtg.obs
+[[ -f fort.68 ]] && mv fort.68 $log_dir/hycom_var.$ddtg.grd
+[[ -f fort.69 ]] && mv fort.69 $log_dir/hycom_var.$ddtg.via
+[[ -f fort.88 ]] && mv fort.88 $log_dir/hycom_var.$ddtg.dbg
 
 #   create data coverage graphics
 DoGraphics=NO
